Removing view model from SortView
