+++ Sources/Models/SortBy.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+enum SortBy {
+    case starsAsc
+    case starsDes
+    case forksAsc
+    case forksDes
+    case updatedAsc
+    case updatedDes
+}
